test(integration): split unit from env/integration tests
