+++ internal/infra/helpers_test.go
@@ -1,0 +1,74 @@
+package infra
+
+import (
+	"testing"
+
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/client/fake"
+
+	openbaov1alpha1 "github.com/dc-tec/openbao-operator/api/v1alpha1"
+)
+
+// testScheme is a shared scheme used across tests.
+var testScheme = func() *runtime.Scheme {
+	scheme := runtime.NewScheme()
+	_ = clientgoscheme.AddToScheme(scheme)
+	_ = openbaov1alpha1.AddToScheme(scheme)
+	return scheme
+}()
+
+func newTestClient(t *testing.T) client.Client {
+	t.Helper()
+
+	// Create the Kubernetes service that NetworkPolicy detection requires
+	kubernetesService := &corev1.Service{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "kubernetes",
+			Namespace: "default",
+		},
+		Spec: corev1.ServiceSpec{
+			ClusterIP: "10.43.0.1", // Used to derive the kubernetes Service IP CIDR (10.43.0.1/32)
+			Ports: []corev1.ServicePort{
+				{
+					Name: "https",
+					Port: 443,
+				},
+			},
+		},
+	}
+
+	return fake.NewClientBuilder().
+		WithScheme(testScheme).
+		WithObjects(kubernetesService).
+		WithReturnManagedFields().
+		Build()
+}
+
+//nolint:unparam // namespace is used in other tests/integration tests
+func newMinimalCluster(name, namespace string) *openbaov1alpha1.OpenBaoCluster {
+	return &openbaov1alpha1.OpenBaoCluster{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      name,
+			Namespace: namespace,
+		},
+		Spec: openbaov1alpha1.OpenBaoClusterSpec{
+			Version:  "2.4.4",
+			Image:    "openbao/openbao:2.4.4",
+			Replicas: 3,
+			TLS: openbaov1alpha1.TLSConfig{
+				Enabled:        true,
+				RotationPeriod: "720h",
+			},
+			Storage: openbaov1alpha1.StorageConfig{
+				Size: "10Gi",
+			},
+			InitContainer: &openbaov1alpha1.InitContainerConfig{
+				Image: "openbao/openbao-init:latest",
+			},
+		},
+	}
+}
